Crate params

Request parameters parser plugin for Iron

This plugin is a multi-source request parameters parser.

Structs

File
An uploaded file that was received as part of multipart/form-data parsing.

Map
A type that maps keys to request parameter values.

Params
Plugin for iron::Request that processes and combines request parameters from the various request sources.

Enums

ParamsError
An error representing any of the possible errors that can occur during parameter processing.

Value
A representation of all possible types of request parameters.

Traits

FromValue
An interface for converting from Value variants.