Struct params::File

pub struct File(_);

An uploaded file that was received as part of multipart/form-data parsing.

Files are streamed to disk because they may not fit in memory.

Methods

impl File

fn open(&self) -> Result<File>

Attempts to open the file in read-only mode.

fn path(&self) -> &Path

The path to the temporary file where the data was saved.

fn filename(&self) -> Option<&str>

The filename that was specified in the request, unfiltered.

Warning

This may or may not be legal on the local filesystem, and so you should not blindly append it to a Path, as such behavior could easily be exploited by a malicious client.

fn size(&self) -> u64

The size of the file, in bytes.