Struct params::Map

pub struct Map(pub BTreeMap<String, Value>);

A type that maps keys to request parameter values.

Methods

impl Map

fn new() -> Map

Creates an empty map.

fn assign(&mut self, path: &str, value: Value) -> Result<(), ParamsError>

Inserts a parameter value to the specified key path.

Key paths are a series of values starting with a plain name and followed by zero or more array-like indices. name is "the value called name", names[] is "the array called names", pts[][x] and pts[][y] are the x and y values of a map at the end of an array.

This method is used during the internal parsing processes and is only made public in the name of hypothetical extensibility.

Examples

let mut map = Map::new();
map.assign("name", Value::String("Callie".into())).unwrap();
assert_eq!(format!("{:?}", map), r#"{"name": "Callie"}"#);

let mut map = Map::new();
map.assign("names[]", Value::String("Anne".into())).unwrap();
map.assign("names[]", Value::String("Bob".into())).unwrap();
assert_eq!(format!("{:?}", map), r#"{"names": ["Anne", "Bob"]}"#);

let mut map = Map::new();
map.assign("pts[][x]", Value::I64(3)).unwrap();
map.assign("pts[][y]", Value::I64(9)).unwrap();
assert_eq!(format!("{:?}", map), r#"{"pts": [{"x": 3, "y": 9}]}"#);

fn find(&self, keys: &[&str]) -> Option<&Value>

Traverses nested Maps to find the specified value by key.

Examples

let mut map = Map::new();
map.assign("user[name]", Value::String("Marie".into())).unwrap();

match map.find(&["user", "name"]) {
    Some(&Value::String(ref name)) => assert_eq!(name, "Marie"),
    _ => panic!("Unexpected parameter type!"),
}

assert!(map.find(&["user", "age"]).is_none());

fn to_strict_map<T: FromValue>(&self) -> Option<BTreeMap<String, T>>

Converts this map to a raw BTreeMap with values of all the same type.

let mut map = Map::new();
map.assign("x", Value::String("10.5".into())).unwrap();
map.assign("y", Value::I64(15)).unwrap();

let mut expected = BTreeMap::new();
expected.insert(String::from("x"), 10.5f32);
expected.insert(String::from("y"), 15.0f32);

assert_eq!(map.to_strict_map::<f32>(), Some(expected));